# server related/javascript-client-generated/src/model/Question.js
/*
 * queriKorner - OpenAPI 3.0
 * This is the API documentation of the queriKorner server application. It is based on the OpenAPI 3.0 specification.
 *
 * OpenAPI spec version: 1.0.11
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 *
 * Swagger Codegen version: 3.0.41
 *
 * Do not edit the class manually.
 *
 */
import {ApiClient} from '../ApiClient';

/**
 * The Question model module.
 * @module model/Question
 * @version 1.0.11
 */
export class Question {
  /**
   * Constructs a new <code>Question</code>.
   * @alias module:model/Question
   * @class
   */
  constructor() {
  }

  /**
   * Constructs a <code>Question</code> from a plain JavaScript object, optionally creating a new instance.
   * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
   * @param {Object} data The plain JavaScript object bearing properties of interest.
   * @param {module:model/Question} obj Optional instance to populate.
   * @return {module:model/Question} The populated <code>Question</code> instance.
   */
  static constructFromObject(data, obj) {
    if (data) {
      obj = obj || new Question();
      if (data.hasOwnProperty('questionId'))
        obj.questionId = ApiClient.convertToType(data['questionId'], 'String');
      if (data.hasOwnProperty('Title'))
        obj.title = ApiClient.convertToType(data['Title'], 'String');
      if (data.hasOwnProperty('Body'))
        obj.body = ApiClient.convertToType(data['Body'], 'String');
      if (data.hasOwnProperty('Tags'))
        obj.tags = ApiClient.convertToType(data['Tags'], ['String']);
      if (data.hasOwnProperty('Votes'))
        obj.votes = ApiClient.convertToType(data['Votes'], 'Number');
      if (data.hasOwnProperty('AnswerIds'))
        obj.answerIds = ApiClient.convertToType(data['AnswerIds'], ['String']);
      if (data.hasOwnProperty('NumberOfAnswers'))
        obj.numberOfAnswers = ApiClient.convertToType(data['NumberOfAnswers'], 'Number');
      if (data.hasOwnProperty('CreatedAt'))
        obj.createdAt = ApiClient.convertToType(data['CreatedAt'], 'String');
      if (data.hasOwnProperty('Accepted'))
        obj.accepted = ApiClient.convertToType(data['Accepted'], 'Boolean');
      if (data.hasOwnProperty('CreatedBy'))
        obj.createdBy = ApiClient.convertToType(data['CreatedBy'], 'String');
    }
    return obj;
  }
}

/**
 * @member {String} questionId
 */
Question.prototype.questionId = undefined;

/**
 * @member {String} title
 */
Question.prototype.title = undefined;

/**
 * @member {String} body
 */
Question.prototype.body = undefined;

/**
 * @member {Array.<String>} tags
 */
Question.prototype.tags = undefined;

/**
 * @member {Number} votes
 */
Question.prototype.votes = undefined;

/**
 * @member {Array.<String>} answerIds
 */
Question.prototype.answerIds = undefined;

/**
 * @member {Number} numberOfAnswers
 */
Question.prototype.numberOfAnswers = undefined;

/**
 * @member {String} createdAt
 */
Question.prototype.createdAt = undefined;

/**
 * @member {Boolean} accepted
 */
Question.prototype.accepted = undefined;

/**
 * @member {String} createdBy
 */
Question.prototype.createdBy = undefined;