# server related/python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: queriKorner - OpenAPI 3.0
  description: This is the API documentation of the queriKorner server application.
    It is based on the OpenAPI 3.0 specification.
  version: 1.0.11
servers:
- url: /
...
paths:
  /question:
    post:
      tags:
      - Question
      summary: Ask your query
      description: Ask your query
      operationId: ask_query
      requestBody:
        description: Query to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
          application/xml:
            schema:
              $ref: '#/components/schemas/Question'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
            application/xml:
              schema:
                $ref: '#/components/schemas/Question'
        "400":
          description: Invalid question
        "404":
          description: Question not found
      security:
      - ask_auth:
        - write:answers
        - read:answers
      x-openapi-router-controller: swagger_server.controllers.question_controller
  /question/{questionId}:
    get:
      tags:
      - Question
      summary: Get all query data
      description: Get all query data
      operationId: get_data
      parameters:
      - name: questionId
        in: query
        description: Query data
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "71932"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
            application/xml:
              schema:
                $ref: '#/components/schemas/Question'
        "400":
          description: Invalid data format
        "404":
          description: Question not found
      security:
      - ask_auth:
        - write:answers
        - read:answers
      x-openapi-router-controller: swagger_server.controllers.question_controller
    put:
      tags:
      - Question
      summary: Change question data
      description: Change question data
      operationId: change_question
      parameters:
      - name: questionId
        in: query
        description: Query data
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "71932"
      requestBody:
        description: Query to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
          application/xml:
            schema:
              $ref: '#/components/schemas/Question'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
            application/xml:
              schema:
                $ref: '#/components/schemas/Question'
        "400":
          description: Invalid data format
        "404":
          description: Question not found
      security:
      - ask_auth:
        - write:answers
        - read:answers
      x-openapi-router-controller: swagger_server.controllers.question_controller
    delete:
      tags:
      - Question
      summary: Delete a query
      description: Delete a query
      operationId: delete_query
      parameters:
      - name: questionId
        in: query
        description: Query data
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "71932"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
            application/xml:
              schema:
                $ref: '#/components/schemas/Question'
        "400":
          description: Invalid data format
        "404":
          description: Question not found
      security:
      - ask_auth:
        - write:answers
        - read:answers
      x-openapi-router-controller: swagger_server.controllers.question_controller
  /question/votes/{questionId}:
    put:
      tags:
      - Question
      summary: Upvote/Downvote a question
      description: Upvote/Downvote a question
      operationId: upvote_question
      parameters:
      - name: questionId
        in: query
        description: Question id
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: "71932"
      requestBody:
        description: Query to be added to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
          application/xml:
            schema:
              $ref: '#/components/schemas/Question'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
            application/xml:
              schema:
                $ref: '#/components/schemas/Question'
        "400":
          description: Invalid data format
        "404":
          description: Question not found
      security:
      - ask_auth:
        - write:answers
        - read:answers
      x-openapi-router-controller: swagger_server.controllers.question_controller
...
components:
  schemas:
    Question:
      type: object
      properties:
        questionId:
          type: string
          example: "71932"
        Title:
          type: string
          example: How to create a new file in Java?
        Body:
          type: string
          example: I am trying to create a new file in Java. I am using the following
            code but it is not working. Can anyone help me?
        Tags:
          type: array
          example:
          - java
          - file
          items:
            type: string
        Votes:
          type: integer
          example: 10
        AnswerIds:
          type: array
          example:
          - "213212"
          - "314132"
          items:
            type: string
        NumberOfAnswers:
          type: integer
          example: 2
        CreatedAt:
          type: string
          example: 2019-10-10T12:00:00Z
        Accepted:
          type: boolean
          example: false
        CreatedBy:
          type: string
          example: Kevin
      example:
        questionId: "71932"
        CreatedBy: Kevin
        CreatedAt: 2019-10-10T12:00:00Z
        AnswerIds:
        - "213212"
        - "314132"
        NumberOfAnswers: 2
        Title: How to create a new file in Java?
        Votes: 10
        Body: I am trying to create a new file in Java. I am using the following code
          but it is not working. Can anyone help me?
        Tags:
        - java
        - file
        Accepted: false
...
  securitySchemes:
    ask_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://queriKorner.swagger.io/oauth/authorize
          scopes:
            write:answers: modify answers in your account
            read:answers: read your answers
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_ask_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_ask_auth
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key